refactor: split tidy CLI into check/fix modules (drop size-waiver)
Replace monofile cmd/tidy.rs with:
- mod.rs: args + thin run dispatch
- check.rs: issue scan and report (tidy check)
- fix.rs: parallel dirty scan, stage, finalize_report (tidy fix)
- tests.rs: co-located unit tests

No behavior change. Reference-doc inventory path updated to tidy/mod.rs.
